requirements.txt
# architect==0.5.6
django==3.0.2
django-cors-headers==3.2.1
django-extensions==2.2.5
django-filter==2.2.0
django-health-check==3.11.1
django-jet==1.0.8
django-leaflet==0.26.0
django-rosetta==0.9.3
django-tgb-basics==0.2.11
djangorestframework==3.11.0
djangorestframework-gis==0.15
djangorestframework-simplejwt==4.4.0
drf-friendly-errors==0.13
drf-yasg==1.17.0
drf-yasg[validation]
gunicorn==20.0.4
psycopg2-binary==2.8.4
python-dotenv==0.10.3
# bandit==1.6.2
# black==19.10b0
django-debug-toolbar==2.1
pylint==2.4.4
pylint-django==2.0.13
